* issue #490

+++ de.bht.fb6.mme2.mfg/module/JumpUpDriver/src/JumpUpDriver/Controller/VehicleController.php
@@ -18,10 +18,12 @@ use Application\Util\FilesUtil;
 /**
  *
  *
+ *
  * This controller handles the addition / modification / removement of user-defined vehicles.
  *
  * @package JumpUpDriver\Controller
  * @subpackage
+ *
  *
  * @copyright Copyright (c) 2013 Sascha Feldmann (http://saschafeldmann.de)
  * @license GNU license
@@ -31,7 +33,7 @@ use Application\Util\FilesUtil;
 class VehicleController extends ANeedsAuthenticationController {
 	/**
 	 * The key determining the value for the parameter of the vehicle.
-	 * 
+	 *
 	 * @var String
 	 */
 	const PARAM_VEHICLE_ID = "vehId";
@@ -63,7 +65,7 @@ class VehicleController extends ANeedsAuthenticationController {
 							'owner' => $user->getId () 
 					) );
 					if (null !== $vehicle) { // show form for editing
-					                        // Create hard-coded inputFields
+					                         // Create hard-coded inputFields
 						$inputFields = array (
 								'<input type="hidden" name="' . self::PARAM_VEHICLE_ID . '" value="' . $vehicle->getId () . '"/>' 
 						);
@@ -90,7 +92,7 @@ class VehicleController extends ANeedsAuthenticationController {
 						$this->flashMessenger ()->addErrorMessage ( IControllerMessages::FATAL_ERROR_NOT_AUTHENTIFICATED );
 						$this->redirect ()->toRoute ( IRouteStore::ADD_TRIP_ERROR );
 					} else { // success & user is logged in
-					       // move pic and get the full path to it
+					         // move pic and get the full path to it
 						if (null !== $data [VehicleForm::FIELD_VEHICLE_PIC]) {
 							$pathVehiclePic = FilesUtil::moveUploadedFile ( $data [VehicleForm::FIELD_VEHICLE_PIC], $user, FilesUtil::TYPE_VEHICLE_PIC, $vehicle );
 							$vehicle->setVehiclepic ( $pathVehiclePic );
@@ -127,16 +129,21 @@ class VehicleController extends ANeedsAuthenticationController {
 							'id' => $vehicleId,
 							'owner' => $user->getId () 
 					) );
-					// remove pic if available
-					$picPath = $vehicle->getVehiclepic ();
-					if (null !== $picPath) {
-						$success = unlink ( $picPath );
+					// check, whether the vehicle is contained in a trip
+					if (! $this->_isContainedInTrip ( $vehicle )) {
+						// remove pic if available
+						$picPath = $vehicle->getVehiclepic ();
+						if (null !== $picPath) {
+							$success = unlink ( $picPath );
+						}
+						$this->em->remove ( $vehicle ); // remove from db
+						$this->em->flush ();
+						$this->flashMessenger ()->clearMessages ();
+						$this->flashMessenger ()->addMessage ( IControllerMessages::SUCCESS_DELETE_VEHICLE );
 					}
-					$this->em->remove ( $vehicle ); // remove from db
-					$this->em->flush ();
-					$this->flashMessenger ()->clearMessages ();
-					$this->flashMessenger ()->addMessage ( IControllerMessages::SUCCESS_DELETE_VEHICLE );
-					
+					else { // vehicle is contained in a trip -> show error message
+						$this->flashMessenger()->addMessage( IControllerMessages::DELETE_VEHICLE_IS_IN_TRIP);
+					}
 				} else {
 					$this->flashMessenger ()->clearMessages ();
 					$this->flashMessenger ()->addMessage ( IControllerMessages::DELETE_VEHICLE_NO_ID );
@@ -176,7 +183,7 @@ class VehicleController extends ANeedsAuthenticationController {
 	 */
 	public function addAction() {
 		if ($this->_checkAuthentication ()) { // auth required
-		                                    // check if the request was redirected by the AddTripController
+		                                      // check if the request was redirected by the AddTripController
 			$infoMessages = "";
 			if ($this->flashMessenger ()->hasInfoMessages ()) {
 				$infoMessages = $this->flashMessenger ()->getInfoMessages ();
@@ -227,5 +234,17 @@ class VehicleController extends ANeedsAuthenticationController {
 					"infoMessages" => $infoMessages 
 			);
 		}
+	}
+	
+	/**
+	 * Check whether the given vehicle is contained in a trip.
+	 * @param Vehicle $vehicle
+	 * @return boolean true if the vehicle is contained in any trip.
+	 */
+	protected function _isContainedInTrip (Vehicle $vehicle) {
+		if(null !== $vehicle->getIntrips() && (sizeof($vehicle->getIntrips() > 0))) {
+			return true;
+		} 
+		return false;
 	}
 }
